ATUALIZACAO_CALCULO-main/requirements.txt
anticaptchaofficial==1.0.59
anyio==4.2.0
attrs==23.2.0
auto-download-undetected-chromedriver==0.16
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
chromedriver-autoinstaller==0.6.3
colorama==0.4.6
downloadunzip==0.10
et_xmlfile==2.0.0
exceptiongroup==1.2.0
FastCopyFast==0.52
flatten-any-dict-iterable-or-whatsoever==0.33
flatten-everything==0.41
getfilenuitkapython==0.11
google-api-core==2.15.0
google-api-python-client==2.165.0
google-auth==2.26.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.62.0
gspread==5.12.4
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.26.0
hyperframe==6.0.1
idna==3.6
list-all-files-recursively==0.12
list-files-with-timestats==0.11
numpy==2.0.2
oauthlib==3.2.2
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.3
protobuf==4.25.2
psutil==6.1.0
py==1.11.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyMySQL==1.1.0
pyparsing==3.1.1
PyPDF2==3.0.1
PySocks==1.7.1
python-anticaptcha==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.2
pywin32==308
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
schedule==1.2.1
selenium==4.16.0
shlexwhichplus==0.10
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
touchtouch==0.11
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2024.2
undetected-chromedriver==3.5.4
uritemplate==4.1.1
urllib3==2.1.0
webdriver-manager==4.0.1
websockets==12.0
wsproto==1.2.0
xlwings==0.33.4
